Move interfaces to Interfaces folder, split SensorDescriptionBuilder business logic to two classes ModuleDescriptionBuilder and DevicesDescriptionBuilder
