+++ src/PxcIolinkConfiguration/Interfaces/IXmlTemplateReader.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using EasyEPlanner.PxcIolinkConfiguration.Models;
 
-namespace EasyEPlanner.PxcIolinkConfiguration
+namespace EasyEPlanner.PxcIolinkConfiguration.Interfaces
 {
     public interface IXmlTemplateReader
     {
