+++ src/PxcIolinkConfiguration/XmlTemplateReader.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using EasyEPlanner.PxcIolinkConfiguration.Interfaces;
 using EasyEPlanner.PxcIolinkConfiguration.Models;
 
 namespace EasyEPlanner.PxcIolinkConfiguration
